# Conflicts:
#	pogom/models.py
#	requirements.txt
#	static/js/map.js

+++ pogom/models.py
@@ -1,16 +1,16 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
-
 import logging
-import os
 import calendar
-import copy
-from peewee import Model, SqliteDatabase, InsertQuery,\
-                   IntegerField, CharField, DoubleField, BooleanField,\
-                   DateTimeField, OperationalError, create_model_tables, fn
+import sys
+import math
+from peewee import SqliteDatabase, InsertQuery, \
+    IntegerField, CharField, DoubleField, BooleanField, \
+    DateTimeField, CompositeKey, fn
 from playhouse.flask_utils import FlaskDB
 from playhouse.pool import PooledMySQLDatabase
 from playhouse.shortcuts import RetryOperationalError
+from playhouse.migrate import migrate, MySQLMigrator, SqliteMigrator
 from datetime import datetime, timedelta
 from base64 import b64encode
 
@@ -25,6 +25,7 @@ args = get_args()
 flaskDb = FlaskDB()
 
 pokemonLured = {}
+db_schema_version = 5
 
 
 class MyRetryDB(RetryOperationalError, PooledMySQLDatabase):
@@ -34,13 +35,16 @@ class MyRetryDB(RetryOperationalError, PooledMySQLDatabase):
 def init_database(app):
     if args.db_type == 'mysql':
         log.info('Connecting to MySQL database on %s:%i', args.db_host, args.db_port)
+        connections = args.db_max_connections
+        if hasattr(args, 'accounts'):
+            connections *= len(args.accounts)
         db = MyRetryDB(
             args.db_name,
             user=args.db_user,
             password=args.db_pass,
             host=args.db_host,
             port=args.db_port,
-            max_connections=args.db_max_connections,
+            max_connections=connections,
             stale_timeout=300)
     else:
         log.info('Connecting to local SQLite database')
@@ -176,7 +180,7 @@ class Pokemon(BaseModel):
         query = (Pokemon
                  .select()
                  .where((Pokemon.pokemon_id == pokemon_id) &
-                        (Pokemon.disappear_time > datetime.utcfromtimestamp(last_appearance/1000.0))
+                        (Pokemon.disappear_time > datetime.utcfromtimestamp(last_appearance / 1000.0))
                         )
                  .order_by(Pokemon.disappear_time.asc())
                  .dicts()
@@ -186,6 +190,78 @@ class Pokemon(BaseModel):
             appearances.append(a)
         return appearances
 
+    @classmethod
+    def get_spawnpoints(cls, swLat, swLng, neLat, neLng):
+        query = Pokemon.select(Pokemon.latitude, Pokemon.longitude, Pokemon.spawnpoint_id)
+
+        if None not in (swLat, swLng, neLat, neLng):
+            query = (query
+                     .where((Pokemon.latitude >= swLat) &
+                            (Pokemon.longitude >= swLng) &
+                            (Pokemon.latitude <= neLat) &
+                            (Pokemon.longitude <= neLng)
+                            )
+                     )
+
+        # Sqlite doesn't support distinct on columns
+        if args.db_type == 'mysql':
+            query = query.distinct(Pokemon.spawnpoint_id)
+        else:
+            query = query.group_by(Pokemon.spawnpoint_id)
+
+        return list(query.dicts())
+
+    @classmethod
+    def get_spawnpoints_in_hex(cls, center, steps):
+        log.info('got {}steps'.format(steps))
+        # work out hex bounding box
+        hdist = ((steps * 120.0) - 50.0) / 1000.0
+        vdist = ((steps * 105.0) - 35.0) / 1000.0
+        R = 6378.1  # km radius of the earth
+        vang = math.degrees(vdist / R)
+        hang = math.degrees(hdist / (R * math.cos(math.radians(center[0]))))
+        north = center[0] + vang
+        south = center[0] - vang
+        east = center[1] + hang
+        west = center[1] - hang
+        # get all spawns in that box
+        query = (Pokemon
+                 .select(Pokemon.latitude.alias('lat'),
+                         Pokemon.longitude.alias('lng'),
+                         ((Pokemon.disappear_time.minute * 60) + (Pokemon.disappear_time.second + 2700) % 3600).alias('time'),
+                         Pokemon.spawnpoint_id
+                         ))
+        query = (query.where((Pokemon.latitude <= north) &
+                             (Pokemon.latitude >= south) &
+                             (Pokemon.longitude >= west) &
+                             (Pokemon.longitude <= east)
+                             ))
+        # Sqlite doesn't support distinct on columns
+        if args.db_type == 'mysql':
+            query = query.distinct(Pokemon.spawnpoint_id)
+        else:
+            query = query.group_by(Pokemon.spawnpoint_id)
+
+        s = list(query.dicts())
+        # for each spawn work out if it is in the hex (clipping the diagonals)
+        trueSpawns = []
+        for spawn in s:
+            # get the offset from the center of each spawn in km
+            offset = [math.radians(spawn['lat'] - center[0]) * R, math.radians(spawn['lng'] - center[1]) * (R * math.cos(math.radians(center[0])))]
+            # check agains the 4 lines that make up the diagonals
+            if (offset[1] + (offset[0] * 0.5)) > hdist:  # too far ne
+                continue
+            if (offset[1] - (offset[0] * 0.5)) > hdist:  # too far se
+                continue
+            if ((offset[0] * 0.5) - offset[1]) > hdist:  # too far nw
+                continue
+            if ((0 - offset[1]) - (offset[0] * 0.5)) > hdist:  # too far sw
+                continue
+            # if it gets to here its  a good spawn
+            trueSpawns.append(spawn)
+        return trueSpawns
+
+
 class Pokestop(BaseModel):
     pokestop_id = CharField(primary_key=True, max_length=50)
     enabled = BooleanField()
@@ -193,7 +269,7 @@ class Pokestop(BaseModel):
     longitude = DoubleField()
     last_modified = DateTimeField(index=True)
     lure_expiration = DateTimeField(null=True, index=True)
-    active_pokemon_id = IntegerField(null=True)
+    active_fort_modifier = CharField(max_length=50, null=True)
 
     class Meta:
         indexes = ((('latitude', 'longitude'), False),)
@@ -264,13 +340,12 @@ class Gym(BaseModel):
 
 
 class ScannedLocation(BaseModel):
-    scanned_id = CharField(primary_key=True, max_length=50)
     latitude = DoubleField()
     longitude = DoubleField()
     last_modified = DateTimeField(index=True)
 
     class Meta:
-        indexes = ((('latitude', 'longitude'), False),)
+        primary_key = CompositeKey('latitude', 'longitude')
 
     @staticmethod
     def get_recent(swLat, swLng, neLat, neLng):
@@ -289,6 +364,14 @@ class ScannedLocation(BaseModel):
             scans.append(s)
 
         return scans
+
+
+class Versions(flaskDb.Model):
+    key = CharField()
+    val = IntegerField()
+
+    class Meta:
+        primary_key = False
 
 
 def parse_map(map_dict, step_location):
@@ -314,9 +397,14 @@ def parse_map(map_dict, step_location):
                         'disappear_time': d_t
                     }
             for p in cell.get('wild_pokemons', []):
-                d_t = datetime.utcfromtimestamp(
-                    (p['last_modified_timestamp_ms'] +
-                     p['time_till_hidden_ms']) / 1000.0)
+                # time_till_hidden_ms was overflowing causing a negative integer. It was also returning a value above 3.6M ms.
+                if (0 < p['time_till_hidden_ms'] < 3600000):
+                    d_t = datetime.utcfromtimestamp(
+                        (p['last_modified_timestamp_ms'] +
+                         p['time_till_hidden_ms']) / 1000.0)
+                else:
+                    # Set a value of 15 minutes because currently its unknown but larger than 15.
+                    d_t = datetime.utcfromtimestamp((p['last_modified_timestamp_ms'] + 900000) / 1000.0)
                 diff = d_t - now
                 if diff.total_seconds() < 1500:
                     pokemons[p['encounter_id']] = {
@@ -421,7 +509,20 @@ def parse_map(map_dict, step_location):
     pokemons_upserted = 0
     pokestops_upserted = 0
     gyms_upserted = 0
-    
+
+    scanned[0] = {
+        'latitude': step_location[0],
+        'longitude': step_location[1],
+        'last_modified': datetime.utcnow(),
+    }
+
+    while True:
+        try:
+            flaskDb.connect_db()
+            break
+        except Exception as e:
+            log.warning('%s... Retrying', e)
+
     if pokemons and config['parse_pokemon']:
         pokemons_upserted = len(pokemons)
         bulk_upsert(Pokemon, pokemons)
@@ -434,33 +535,33 @@ def parse_map(map_dict, step_location):
         gyms_upserted = len(gyms)
         bulk_upsert(Gym, gyms)
 
-    log.info('Upserted %d pokemon, %d pokestops, and %d gyms',
-        pokemons_upserted,
-        pokestops_upserted,
-        gyms_upserted)
-
-    scanned[0] = {
-        'scanned_id': str(step_location[0])+','+str(step_location[1]),
-        'latitude': step_location[0],
-        'longitude': step_location[1],
-        'last_modified': datetime.utcnow(),
-    }
-
     bulk_upsert(ScannedLocation, scanned)
 
     clean_database()
+
+    flaskDb.close_db(None)
+
+    log.info('Upserted %d pokemon, %d pokestops, and %d gyms',
+             pokemons_upserted,
+             pokestops_upserted,
+             gyms_upserted)
 
     return True
 
 
 def clean_database():
-    flaskDb.connect_db()
     query = (ScannedLocation
-            .delete()
-            .where((ScannedLocation.last_modified <
-                (datetime.utcnow() - timedelta(minutes=30)))))
+             .delete()
+             .where((ScannedLocation.last_modified <
+                    (datetime.utcnow() - timedelta(minutes=30)))))
     query.execute()
-    flaskDb.close_db(None)
+
+    if args.purge_data > 0:
+        query = (Pokemon
+                 .delete()
+                 .where((Pokemon.disappear_time <
+                        (datetime.utcnow() - timedelta(hours=args.purge_data)))))
+        query.execute()
 
 
 def bulk_upsert(cls, data):
@@ -468,27 +569,91 @@ def bulk_upsert(cls, data):
     i = 0
     step = 120
 
-    flaskDb.connect_db()
-
     while i < num_rows:
-        log.debug('Inserting items %d to %d', i, min(i+step, num_rows))
+        log.debug('Inserting items %d to %d', i, min(i + step, num_rows))
         try:
-            InsertQuery(cls, rows=data.values()[i:min(i+step, num_rows)]).upsert().execute()
+            InsertQuery(cls, rows=data.values()[i:min(i + step, num_rows)]).upsert().execute()
         except Exception as e:
             log.warning('%s... Retrying', e)
             continue
 
-        i+=step
-
-    flaskDb.close_db(None)
+        i += step
 
 
 def create_tables(db):
     db.connect()
+    verify_database_schema(db)
     db.create_tables([Pokemon, Pokestop, Gym, ScannedLocation], safe=True)
     db.close()
 
+
 def drop_tables(db):
     db.connect()
-    db.drop_tables([Pokemon, Pokestop, Gym, ScannedLocation], safe=True)
+    db.drop_tables([Pokemon, Pokestop, Gym, ScannedLocation, Versions], safe=True)
     db.close()
+
+
+def verify_database_schema(db):
+    if not Versions.table_exists():
+        db.create_tables([Versions])
+
+        if ScannedLocation.table_exists():
+            # Versions table didn't exist, but there were tables. This must mean the user
+            # is coming from a database that existed before we started tracking the schema
+            # version. Perform a full upgrade.
+            InsertQuery(Versions, {Versions.key: 'schema_version', Versions.val: 0}).execute()
+            database_migrate(db, 0)
+        else:
+            InsertQuery(Versions, {Versions.key: 'schema_version', Versions.val: db_schema_version}).execute()
+
+    else:
+        db_ver = Versions.get(Versions.key == 'schema_version').val
+
+        if db_ver < db_schema_version:
+            database_migrate(db, db_ver)
+
+        elif db_ver > db_schema_version:
+            log.error("Your database version (%i) appears to be newer than the code supports (%i).",
+                      db_ver, db_schema_version)
+            log.error("Please upgrade your code base or drop all tables in your database.")
+            sys.exit(1)
+
+
+def database_migrate(db, old_ver):
+    # Update database schema version
+    Versions.update(val=db_schema_version).where(Versions.key == 'schema_version').execute()
+
+    log.info("Detected database version %i, updating to %i", old_ver, db_schema_version)
+
+    # Perform migrations here
+    migrator = None
+    if args.db_type == 'mysql':
+        migrator = MySQLMigrator(db)
+    else:
+        migrator = SqliteMigrator(db)
+
+#   No longer necessary, we're doing this at schema 4 as well
+#    if old_ver < 1:
+#        db.drop_tables([ScannedLocation])
+
+    if old_ver < 2:
+        migrate(migrator.add_column('pokestop', 'encounter_id', CharField(max_length=50, null=True)))
+
+    if old_ver < 3:
+        migrate(
+            migrator.add_column('pokestop', 'active_fort_modifier', CharField(max_length=50, null=True)),
+            migrator.drop_column('pokestop', 'encounter_id'),
+            migrator.drop_column('pokestop', 'active_pokemon_id')
+        )
+
+    if old_ver < 4:
+        db.drop_tables([ScannedLocation])
+
+    if old_ver < 5:
+        # Some pokemon were added before the 595 bug was "fixed"
+        # Clean those up for a better UX
+        query = (Pokemon
+                 .delete()
+                 .where(Pokemon.disappear_time >
+                        (datetime.utcnow() - timedelta(hours=24))))
+        query.execute()
